Huge cleanup of logging implementation.
* Boost.Log is no longer optional: it *is* our logger.
  - Reflect-in/cleanup CMake
* The implementation was unnecessarily spread across 7 files.
  We now only need 2. Because the implementation is small and
  restricted, opaque pointers do not need their respective classes
  in separate files. We can now clearly see what we are doing and
  how we can do it better.
* Cherry pick most of 8605937 and 88a511e
* Refactor kovri::log -> i2p::util::log
  The kovri namespace is the right direction, but we need to
  do a massive refactor to reflect that instead of patchwork,
  otherwise it looks ugly and confusing (especially in Doxygen).
* This commit brings us closer to resolving #33.
